requirements.txt
Django==3.2.8
djangorestframework
djangorestframework-simplejwt
